+++ yolo_model_development_kit/performance_evaluation_pipeline/source/__init__.py
@@ -1,6 +1,6 @@
 # source/__init__.py
 
-from .dataset_utils import process_labels
+from .dataset_utils import log_label_counts, process_labels
 from .plot_utils import save_fscore_curve, save_pr_curve
 from .run_custom_coco_eval import execute_custom_coco_eval
 from .yolo_evaluation import YoloEvaluator
